I'm making an exception for the word "copyright" which does not appear in the Oxford Advanced Learner's Dictionary. However, it is a very meaningful word in the context of a software licnese, and I suspect any changes to reword it would be harmful. Frankly, if the person reading this license does not understand what copyright is, then nothing I can do to improve the wording will help.
Since copyright is excepted, does that mean I'll revert the changing of
"copyright holder" to "author"? No. Firstly, "holder" would have to be
excepted as well, and there's not a compelling reason to do that.
Secondly, I believe "author" to be a much better term in this context
than "copyright holder" anyway.