Guarantee is used in practice for warranty
http://www.oxfordlearnersdictionaries.com/us/definition/english/guarantee_1
Definition 2:

"a written promise given by a company that something you buy will be replaced or repaired without payment if it goes wrong within a particular period. Synonym warranty.
We provide a 5-year guarantee against rust."

Personally, I'm skeptical of the changes in the warranty disclaimer... but at least this one word should be fine.